+++ pyrekordbox/config.py
@@ -8,23 +8,15 @@ Contains all the path and settings handling of the Rekordbox installation(s) on
 the users machine.
 """
 
-import base64
 import json
 import logging
 import os
-import re
 import sys
-import textwrap
-import time
 import xml.etree.cElementTree as xml
 from pathlib import Path
 from typing import Any, Dict, List, Union
 
-import blowfish
-import frida
 import packaging.version
-
-from .utils import get_rekordbox_pid
 
 logger = logging.getLogger(__name__)
 
@@ -65,10 +57,6 @@ def get_appdata_dir() -> Path:
         logger.warning(f"OS {sys.platform} not supported!")
         return Path("~").expanduser() / ".local" / "share"
     return app_data
-
-
-def get_cache_file() -> Path:
-    return get_appdata_dir() / "pyrekordbox" / _cache_file_name
 
 
 def get_pioneer_install_dir(path: Union[str, Path] = None) -> Path:  # pragma: no cover
@@ -231,46 +219,6 @@ def read_rekordbox6_options(pioneer_app_dir: Union[str, Path]) -> Dict[str, Any]
     return options
 
 
-def read_rekordbox6_asar(rb6_install_dir: Union[str, Path]) -> str:
-    """Finds and parses the Rekordbox 6 `app.asar` archive file.
-
-    An ASAR file is an archive used to package source code for an application
-    using Electron. Rekordbox 6 stores some useful information for opening the
-    new `master.db` database in this file.
-
-    Parameters
-    ----------
-    rb6_install_dir : str or Path
-        The path of the Rekordbox 6 installation directory.
-
-    Returns
-    -------
-    asar_data : str
-        The data of the Rekordbox 6 `app.asar` archive file as ANSI encoded string.
-    """
-    rb6_install_dir = Path(rb6_install_dir)
-    # Find location of app asar file
-    if sys.platform == "win32":
-        location = rb6_install_dir / "rekordboxAgent-win32-x64" / "resources"
-        encoding = "ANSI"
-    elif sys.platform == "darwin":
-        if not str(rb6_install_dir).endswith(".app"):
-            rb6_install_dir = rb6_install_dir / "rekordbox.app"
-        location = (
-            rb6_install_dir / "Contents" / "MacOS" / "rekordboxAgent.app" / "Contents" / "Resources"
-        )
-        encoding = "cp437"
-    else:
-        logger.warning(f"OS {sys.platform} not supported!")
-        return ""
-
-    # Read asar file
-    path = (location / "app.asar").absolute()
-    with open(path, "rb") as fh:
-        data = fh.read().decode(encoding, errors="replace")
-    return data
-
-
 def _extract_version(name: str, major_version: int) -> str:
     name = name.replace(".app", "")  # Needed for MacOS
     ver_str = name.replace("rekordbox", "").strip()
@@ -387,199 +335,6 @@ def _get_rb5_config(
     return conf
 
 
-def _extract_pw(pioneer_install_dir: Path) -> str:  # pragma: no cover
-    """Extract the password for decrypting the Rekordbox 6 database key."""
-    asar_data = read_rekordbox6_asar(pioneer_install_dir)
-    match_result = re.search('pass: ".(.*?)"', asar_data)
-    if match_result is None:
-        raise RuntimeError("Could not read `app.asar` file.")
-    match = match_result.group(0)
-    pw = match.replace("pass: ", "").strip('"')
-    return pw
-
-
-class KeyExtractor:
-    """Extracts the Rekordbox database key using code injection.
-
-    This method works by injecting code into the Rekordbox process and intercepting the
-    call to unlock the database. Works for any Rekordbox version.
-    """
-
-    # fmt: off
-    SCRIPT = textwrap.dedent("""
-    var sqlite3_key = Module.findExportByName(null, 'sqlite3_key');
-
-    Interceptor.attach(sqlite3_key, {
-        onEnter: function(args) {
-            var size = args[2].toInt32();
-            var key = args[1].readUtf8String(size);
-            send('sqlite3_key: ' + key);
-        }
-    });
-    """)
-    # fmt: on
-
-    def __init__(self, rekordbox_executable: Union[str, Path]):
-        self.executable = str(rekordbox_executable)
-        self.key = ""
-
-    def on_message(self, message: Dict[str, Any], data: Any) -> None:
-        payload = message["payload"]
-        if payload.startswith("sqlite3_key"):
-            self.key = payload.split(": ")[1]
-
-    def run(self) -> str:
-        pid = get_rekordbox_pid()
-        if pid:
-            raise RuntimeError(
-                "Rekordbox is running. Please close Rekordbox before running the `KeyExtractor`."
-            )
-        # Spawn Rekordbox process and attach to it
-        pid = frida.spawn(self.executable)
-        frida.resume(pid)
-        session = frida.attach(pid)
-        script = session.create_script(self.SCRIPT)
-        script.on("message", self.on_message)
-        script.load()
-        # Wait for key to be extracted
-        while not self.key:
-            time.sleep(0.1)
-        # Kill Rekordbox process
-        frida.kill(pid)
-
-        return self.key
-
-
-def write_db6_key_cache(key: str) -> None:  # pragma: no cover
-    r"""Writes the decrypted Rekordbox6 database key to the cache file.
-
-    This method can also be used to manually cache the database key, provided
-    the user has found the key somewhere else. The key can be, for example,
-    found in some other projects that hard-coded it.
-
-    The cache file is stored in the application data directory of pyrekordbox:
-    Windows: `C:\Users\<user>\AppData\Roaming\pyrekordbox`
-    macOS: `~/Library/Application Support/pyrekordbox`
-
-    Parameters
-    ----------
-    key : str
-        The decrypted database key. To make sure the key is valid, the first
-        five characters are checked before writing the key to the cache file.
-        The key should start with '402fd'.
-
-    Examples
-    --------
-    >>> from pyrekordbox.config import write_db6_key_cache
-    >>> from pyrekordbox import Rekordbox6Database
-    >>> write_db6_key_cache("402fd...")
-    >>> db = Rekordbox6Database()  # The db can now be opened without providing the key
-    """
-    # Check if the key looks like a valid key
-    if not key.startswith("402fd"):
-        raise ValueError("The provided database key doesn't look valid!")
-    lines = list()
-    lines.append(f"version: {_cache_file_version}")
-    lines.append("dp: " + key)
-    text = "\n".join(lines)
-
-    cache_file = get_cache_file()
-    if not cache_file.parent.exists():
-        cache_file.parent.mkdir()
-
-    with open(cache_file, "w") as fh:
-        fh.write(text)
-    # Set the config key to make sure the key is present after calling method
-    if __config__["rekordbox6"]:
-        __config__["rekordbox6"]["dp"] = key  # type: ignore
-    if __config__["rekordbox7"]:
-        __config__["rekordbox7"]["dp"] = key  # type: ignore
-
-
-def _update_sqlite_key(opts: Dict[str, Any], conf: Dict[str, Any]) -> None:
-    cache_version = 0
-    pw, dp = "", ""
-
-    cache_file = get_cache_file()
-
-    if cache_file.exists():  # pragma: no cover
-        logger.debug("Found cache file %s", cache_file)
-        # Read cache file
-        with open(cache_file, "r") as fh:
-            text = fh.read()
-        lines = text.splitlines()
-        if lines[0].startswith("version:"):
-            cache_version = int(lines[0].split(":")[1].strip())
-        else:
-            cache_version = 1
-        if cache_version == 1:
-            # Cache file introduced in pyrekordbox 0.1.6 contains only the password
-            pw = lines[0]
-            logger.debug("Found pw in cache file")
-        elif cache_version == 2:
-            # Cache file introduced in pyrekordbox 0.1.7 contains version and db key
-            dp = lines[1].split(":")[1].strip()
-            logger.debug("Found dp in cache file")
-        else:
-            raise ValueError(f"Invalid cache version: {cache_version}")
-    else:
-        logger.debug("No cache file found")
-
-    if cache_version < _cache_file_version:  # pragma: no cover
-        # Update cache file
-        if not pw:
-            logger.debug("Extracting pw")
-            try:
-                pw = _extract_pw(conf["install_dir"])
-                logger.debug("Extracted pw from 'app.asar'")
-            except (FileNotFoundError, RuntimeError):
-                logger.debug("Could not extract pw from 'app.asar'")
-                pw = ""
-
-        if not dp:
-            if pw:
-                cipher = blowfish.Cipher(pw.encode())
-                dp = base64.standard_b64decode(opts["dp"])  # type: ignore
-                dp = b"".join(cipher.decrypt_ecb(dp)).decode()
-                logger.debug("Unlocked dp from pw: %s", dp)
-            else:
-                if sys.platform == "win32":
-                    executable = conf["install_dir"] / "rekordbox.exe"
-                elif sys.platform == "darwin":
-                    install_dir = conf["install_dir"]
-                    if not str(install_dir).endswith(".app"):
-                        install_dir = install_dir / "rekordbox.app"
-                    executable = install_dir / "Contents" / "MacOS" / "rekordbox"
-                else:
-                    # Linux: not supported
-                    logger.warning(f"OS {sys.platform} not supported!")
-                    executable = Path("")
-
-                if not executable.exists():
-                    logger.warning(f"Could not find Rekordbox executable: {executable}")
-                else:
-                    extractor = KeyExtractor(executable)
-                    try:
-                        dp = extractor.run()
-                        logger.debug("Extracted dp from Rekordbox process")
-                    except Exception as e:
-                        logger.info(f"`KeyExtractor` failed: {e}")
-
-        if dp:
-            logger.debug("Writing dp to cache file")
-            write_db6_key_cache(dp)
-        else:
-            logging.warning(
-                "Could not retrieve db-key or read it from cache file,"
-                "use the CLI to download it from external sources: "
-                "`python -m pyrekordbox download-key`"
-            )
-
-    # Add database key to config if found
-    if dp:
-        conf["dp"] = dp
-
-
 def _get_rb6_config(
     pioneer_prog_dir: Path, pioneer_app_dir: Path, dirname: str = ""
 ) -> Dict[str, Any]:
@@ -593,9 +348,6 @@ def _get_rb6_config(
     db_dir = db_path.parent
     assert str(conf["db_dir"]) == str(db_dir)
     assert str(conf["db_path"]) == str(db_path)
-
-    # Update SQLite key
-    _update_sqlite_key(opts, conf)
 
     return conf
 
@@ -614,67 +366,7 @@ def _get_rb7_config(
     assert str(conf["db_dir"]) == str(db_dir)
     assert str(conf["db_path"]) == str(db_path)
 
-    # Update SQLite key
-    _update_sqlite_key(opts, conf)
-
     return conf
-
-
-# noinspection PyPackageRequirements,PyUnresolvedReferences
-def _read_config_file(path: Union[str, Path]) -> Dict[str, Any]:
-    path = Path(path)
-    if not path.exists():
-        raise FileNotFoundError(f"No such file or directory: '{path}'")
-
-    ext = path.suffix.lower()
-    if ext == ".cfg":
-        from configparser import ConfigParser
-
-        parser = ConfigParser()
-        parser.read(path)
-        return {k: v for k, v in parser.items() if v is not None}
-    elif ext == ".toml":
-        import toml
-
-        return dict(toml.load(path))
-    elif ext in (".yaml", ".yml"):
-        import yaml
-
-        with open(path, "r") as stream:
-            return dict(yaml.safe_load(stream))
-    return dict()
-
-
-def read_pyrekordbox_configuration() -> Dict[str, Any]:
-    """Reads the pyrekordbox configuration.
-
-    So far only the `pioneer-install-dir` and `pioneer-app-dir` fileds in the
-    `rekordbox` section are supported. Supported configuration files are
-    pyproject.toml, setup.cfg, rekordbox.toml, rekordbox.cfg and rekordbox.yml.
-    The files are searched for in that order.
-
-    Returns
-    -------
-    pyrekordbox_config : dict
-        The pyrekordbox configuration data, if found.
-    """
-    files = ["pyproject.toml", "setup.cfg"]
-    for ext in [".toml", ".cfg", ".yaml", ".yml"]:
-        files.append("pyrekordbox" + ext)
-
-    for file in files:
-        try:
-            data = _read_config_file(file)
-            config = dict(data["rekordbox"])
-            logger.debug("Read configuration from '%s'", file)
-        except (ImportError, FileNotFoundError) as e:
-            logger.debug("Could not read config file '%s': %s", file, e)
-        except KeyError:
-            pass
-        else:
-            if config:
-                return config
-    return dict()
 
 
 def update_config(
@@ -714,7 +406,7 @@ def update_config(
         directory name is used (Windows: 'rekordbox 7.x.x', macOS: 'rekordbox 7.app').
     """
     # Read config file
-    conf = read_pyrekordbox_configuration()
+    conf = dict()
     if pioneer_install_dir is None and "pioneer-install-dir" in conf:
         pioneer_install_dir = conf["pioneer-install-dir"]
     if pioneer_app_dir is None and "pioneer-app-dir" in conf: